+++ src/core_modules/capture-core/components/Pages/MainPage/TeiWorkingLists/epics/index.js
@@ -1,0 +1,2 @@
+// @flow
+export { initTeiViewEpic, updateTeiListEpic } from './teiViewEpics';